+++ extensions/default/src/commandsModule.ts
@@ -150,7 +150,7 @@ const commandsModule = ({
       });
     },
     clearMeasurements: () => {
-      measurementService.clear();
+      measurementService.clearMeasurements();
     },
 
     /**