+++ platform/ui-next/src/components/Icons/Sources/Pencil.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import type { IconProps } from '../types';
+
+export const Pencil = (props: IconProps) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 28 28"
+  >
+    <path
+      fill="currentColor"
+      d="M1.439 16.873l-1.439 7.127 7.128-1.437 16.873-16.872-5.69-5.69-16.872 16.872zm4.702 3.848l-3.582.724.721-3.584 2.861 2.86zm15.031-15.032l-13.617 13.618-2.86-2.861 10.825-10.826 2.846 2.846 1.414-1.414-2.846-2.846 1.377-1.377 2.861 2.86z"
+    />
+  </svg>
+);
+
+export default Pencil;